--- src/sections/dashboard/ChartWrapper.tsx
@@ -1,148 +1,0 @@
-import React from 'react';
-import { XMarkIcon } from '@heroicons/react/20/solid';
-import {
-  AreaChart,
-  BarChart,
-  DonutChart,
-  Legend,
-  LineChart,
-} from '@tremor/react';
-
-import { useMy } from '@/hooks';
-import { Account, ChartType, MyResourceKey, Risk } from '@/types';
-import {
-  CountData,
-  getAggregates as getAccountAggregates,
-  runAggregate as runAccountAggregate,
-} from '@/utils/aggregates/account';
-import {
-  getAggregates as getRiskAggregates,
-  runAggregate as runRiskAggregate,
-} from '@/utils/aggregates/risk';
-
-interface ChartProps {
-  type: ChartType;
-  data: CountData[];
-  xField: string;
-  yField: string;
-}
-
-const Chart: React.FC<ChartProps> = ({ type, data, xField, yField }) => {
-  switch (type) {
-    case 'area':
-      return (
-        <AreaChart
-          className="size-full !h-full"
-          data={data}
-          index={xField}
-          categories={[yField]}
-        />
-      );
-    case 'bar':
-      return (
-        <BarChart
-          className="size-full !h-full"
-          data={data}
-          index={xField}
-          categories={[yField]}
-          colors={['blue']}
-        />
-      );
-    case 'line':
-      return (
-        <LineChart
-          className="size-full !h-full"
-          data={data}
-          index={xField}
-          categories={[yField]}
-          colors={['indigo', 'rose']}
-        />
-      );
-    case 'donut':
-      return (
-        <div className="flex items-center justify-center space-x-6 text-default">
-          <DonutChart
-            className="size-full"
-            variant="donut"
-            data={data}
-            index={xField}
-            category={yField}
-          />
-          <Legend categories={data.map(d => d[xField].toString())} />
-        </div>
-      );
-    default:
-      return null;
-  }
-};
-
-interface ChartWrapperProps {
-  i: string;
-  type: ChartType;
-  endpoint: MyResourceKey;
-  aggregate: string;
-  removeChart: (chartId: string) => void;
-}
-
-const ChartWrapper: React.FC<ChartWrapperProps> = ({
-  i,
-  type,
-  endpoint,
-  aggregate,
-  removeChart,
-}) => {
-  const { data, isLoading } = useMy({
-    resource: endpoint,
-    query: '',
-  });
-
-  if (!isLoading && data) {
-    const getAggregates = () => {
-      switch (endpoint) {
-        case 'account':
-          return getAccountAggregates();
-        case 'risk':
-          return getRiskAggregates();
-        default:
-          return {};
-      }
-    };
-    const runAggregate = (aggregate: string, data: Account[] | Risk[]) => {
-      switch (endpoint) {
-        case 'account':
-          return runAccountAggregate(aggregate, data as Account[]);
-        case 'risk':
-          return runRiskAggregate(aggregate, data as Risk[]);
-        default:
-          return [];
-      }
-    };
-    const aggregateFunction = getAggregates()[aggregate];
-    const chartData = runAggregate(aggregate, data as Account[]);
-
-    return (
-      <div className="relative flex size-full flex-col border border-gray-200 bg-white p-4">
-        <h3 className="font-semibold">{aggregateFunction.label}</h3>
-        <Chart
-          type={type}
-          data={chartData}
-          xField={aggregateFunction.xField}
-          yField={aggregateFunction.yField}
-        />
-        <button
-          onClick={() => removeChart(i)}
-          className="absolute -right-[22px] -top-[22px] m-2 size-6 rotate-45 rounded-full border border-gray-200 bg-white text-gray-500"
-          style={{
-            borderBottomColor: 'transparent',
-          }}
-        >
-          <XMarkIcon className="rotate-45" />
-        </button>
-      </div>
-    );
-  } else {
-    return <div>Loading...</div>; // Handle loading state
-  }
-};
-
-export default ChartWrapper;
